requirements.txt
groq
pillow
transformers
torch
google-colabs